Create PointController and move day event logic
